acceptance-tests/requirements.txt
Appium-Python-Client==0.26
py==1.5.2
pytest==2.9.0
selenium==3.8.1
